+++ test/manifest.js
@@ -39,3 +39,4 @@ require('./test_map');
 require('./test_flatmap');
 require('./test_drop');
 require('./test_inheritance');
+require('./test_and');