+++ option-t.d.ts
@@ -41,6 +41,7 @@ declare module 'option-t' {
         map<U>(fn: (v: T) => U): Option<U>;
         flatMap<U>(fn: (v: T) => Option<U>): Option<U>;
         and<U>(optb: Option<U>): Option<U>;
+        or(optb: Option<T>): Option<T>;
         drop(): void;
     }
 
@@ -51,6 +52,7 @@ declare module 'option-t' {
         map<U>(fn: (v: T) => U): Option<U>;
         flatMap<U>(fn: (v: T) => Option<U>): Option<U>;
         and<U>(optb: Option<U>): Option<U>;
+        or(optb: Option<T>): Option<T>;
         drop(): void;
     }
 
@@ -61,6 +63,7 @@ declare module 'option-t' {
         map<U>(fn: (v: T) => U): Option<U>;
         flatMap<U>(fn: (v: T) => Option<U>): Option<U>;
         and<U>(optb: Option<U>): Option<U>;
+        or(optb: Option<T>): Option<T>;
         drop(): void;
     }
 }