+++ test/manifest.js
@@ -40,3 +40,4 @@ require('./test_flatmap');
 require('./test_drop');
 require('./test_inheritance');
 require('./test_and');
+require('./test_or');